[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14268 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,9 +7,9 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringBuilder sb = new StringBuilder();
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int n = Integer.parseInt(st.nextToken());
-		int m = Integer.parseInt(st.nextToken());
-		int result = (n + m) * (n - m);
+		long n = Long.parseLong(st.nextToken());
+		long m = Long.parseLong(st.nextToken());
+		long result = (n + m) * (n - m);
 		sb.append(result);
 		System.out.println(sb);
 	}
